WL14212: (XProtocol) Support connection close notification
The Purpose of this worklog is to collect and use the X Plugin notices send on
Mysqlx.Notice.Warning messages when it closes the client connection due to one
of the following reasons:

 - The connection is idle (Mysqlx_read_timeout and/or Mysqlx_wait_timeout have
   been exceeded).
 - The connection has been killed by a different session (either via SQL or via
   Admin API).
 - The server is shutting down.

This worklog advices the user why her connection has been closed by a logging
warning and a error message with one of the above reasons. The connection pool
will avoid further connections to this server for a small period of time if the
reason is that the server is shutting down.