BUG32496788: Prepared statements accepts any type of parameters
For prepared statements any type or argument were possible to be used,
which could produce undesired results. This patch enforces the use of
list or type objects for the argument and If the parameter for the
prepared statement is not of type list or tuple an error will be raised.