WL14237: Support Query Attributes
This worklog defines how Connector Python Classic protocol will support
and implement query attributes that apply to statements sent to the
server for execution and available on MySQL v8.0.25 and up. The query
attributes are additional fragments of data that consists of a value,
a type of value and a name that can be used to access this value on
the server's side.