BUG32778827: Raise an error if the _id is different when replacing a document
If a document provided for replacement, using the Collection.replace_one()
or Collection.add_or_replace_one() methods, contains an "_id" different
from the original document, the new "_id" is ignored.

This patch fixes this error by raising an error if the "_id" in the
document provided for replacement is different from the original.

Tests were added for regression.