WL14634: Running unit tests against external server
Currently, unit tests are written so that the testing framework starts
servers required by the tests.

This worklog aims to modify existing unit tests so that they can also
run against an external server not controlled by the testing framework.

The following command line arguments can be used:

  --use-external-server  Sets to use an external server.
                         (Default: False)
  --host                 Hostname or IP address for TCP/IP connections
                         to use with an external server.
                         (Default: 127.0.0.1)
  --port                 TCP/IP port to use with an external server.
                         (Default: 8806)
  --mysqlx-port          MySQL X TCP/IP port to use with an external
                         server.
                         (Default: 88060)
  --user                 User to use with an external server.
                         (Default: root)
  --password             Password to with an external server.
                         (Default: "")

Example:

  python unittest.py --use-external-server --host=10.0.0.32

As an alternative to command line arguments, environment variables can
be used instead:

  host:        $MYSQL_HOST
  port:        $MYSQL_PORT
  mysqlx port: $MYSQLX_PORT
  user:        $MYSQL_USER
  password:    $MYSQL_PASSWORD

The testing system will use the command line arguments, but if they are
not set, the system will try to use the environment variables. Finally,
if no arguments or environment are set, the MySQL defaults will be used.