BUG31528783: Fix number with ZEROFILL not handled by C extension
An exception is raised when a SELECT statement fetch an INT ZEROFILL
data type that will be left 0-padded.

This patch fixes this issue by checking if the ZEROFILL_FLAG flag is
set and passing the base 10 in the PyLong_FromString call used for the
conversion.

A test was added for regression.