BUG32947160: Remove MySQLdb module dependency from Django backend
MySQL Connector/Python extends the Django built-in MySQL backend, which
imports MySQLdb in base.py and introspection.py modules, and if it's not
installed an exception is raised mentioning that the module is missing.

This patch removes the need of using the built-ini MySQL base classes in
base.py and introspection.py, by implementing the missing methods.