WL14689: Fallback conversion to str for types incompatible with MySQL
Currently, if one tries to insert a value of some type that is not
compatible with the default Connector/Python converter, an error is
raised.

This patch introduces a new connection option 'converter_str_fallback',
that allows the conversion to str of values types not supported by the
Connector/Python default converter class, if set to True.

In addition, this patch enables the usage of a custom class when the
C extension is used.