WL#14860: Support FIDO authentication (c-ext)
This worklog implements the FIDO U2F authentication, which is available
since MySQL 8.0.27.

It allows users to create password-less single or multi-factor
authentication accounts that can connect to a MySQL server without using
a password for the corresponding authenticator factor.

This functionality is only available in the C extension implementation.