BUG#21528553: Fix API inconsistency when using consume_results=True
When using connections with `consume_results=True` in the C extension
implementation that have pending results to be fetch, some methods such
as cmd_refresh(), reset_session() and commit(), throws an error.

With this patch, the unread results are handled and consumed if the
`consume_results` connection option is True.

Tests added for regression.