BUG#33203161: Exception is thrown on close connection with pooling
Servers that suffers from Bug#32391415 fixed on 5.7.35, will fail to execute
change_user command due to the default COLLATION used on C/Py. This command
was used in some situations while pooling, that causes to raise an exception
while closing a pooled connection. This patch fixes the by reconnecting on
those situations. In addition the reset command was change on pure to mimic
the c-ext behavior which no raises an error if the command is not recognized
by the server, and just return a bool value True for success and false for
a unsuccess execution.