BUG#33923516: Allow tuple of dictionaries as "failover" argument
In Connector/Python documentation states that the connect() method
accepts a failover argument that provides information to use for the
server failover in the event of connection failures, and this argument
value should be a tuple or a list of dictionaries.

Using a tuple fails with an AttributeError exception.

This patch fixes this issue by applying the sorted() function instead
of using the sort() method in the failover argument, for sorting the
servers by priority.

Change-Id: I8d6911d70d425317374539a830a839ca1cff751f