BUG#28491115: Connector/Python crashes on 0 time value
The Connector/Python binary protocol implemented in
MySQLProtocol._parse_binary_time() crashes on time value of 0
because it does not deal with payload of size 0 correctly.

Expected behavior is to receive a 00:00:00 time when using a
0-time value payload. With this patch, the issue is fixed by
addressing the edge case in which packet size is 0.

This patch also fixes BUG#34006512 (Connector/Python Throws
Exception When Time Field Is 0).

Change-Id: If5efa22812215fb90c86589aed5ca1084b79723f