BUG#27634910: Add warning count method to cursors
The Connector/Python cursor class has a variable to track
the number of warnings generated by the previously
executed operation, but it is private.

With this patch, a new warning_count property is included
in the abstract cursor class, it is publicly available and
can be used to retrieve the number of warnings generated
by the previously executed operation.

Change-Id: I7f6047269cf16e372d79880dcb2a24b7f1440b64