BUG#21402805: Unbound local error when charset name is given as empty to set_charset_collation()
The MySQLConnection.set_charset_collation() method is returning an
UnboundLocalError exception when charset name is given as empty
(it could be either None, 0 or an empty string).

With this patch, this behavior is corrected. Now, every time an
empty data type is assigned to the charset variable, the default
charset is used automatically. Additionally, new instructions were
added to guarantee the charset variable is either int, str, or None,
else a ValueError() exception arises.

The flow described above also applies to the collation variable
with the difference that this latter can be of type either str
or None, but not int.

The bug described so far was present for both connectors
"pure python" and "c-ext".

Furthermore, another issue only present on the "pure python" version
was corrected with this patch. When using the cmd_change_user() method
and providing a negative or non-integer charset value, a struct.error
exception appears.

It was corrected by incorporating an exception for non-integer values,
and for integer values that are not zero or positive.

Change-Id: I6ccdac9bdf890d7daf8fd8a0b2e16829cb7e7bc2