BUG#21463298: Fix weakly-referenced object no longer exists exception
Connector/Python cursors contain a weak reference to the connection,
and a weak reference to an object is not enough to keep the object
alive. So when the connection object is destroyed and the cursor is
used, a ReferenceError exception is raised saying that the
weakly-referenced object no longer exists.

With this patch a ProgrammingError exception is raised instead saying
the cursor is not connected.

Change-Id: I843a9e84185f8b8710383dd18d9df54910bc5840