BUG#21529893: Resultset handling not proper in C-Python with c-ext
The bug report describes the finding of two issues happening when
using the Connector/Python c-extension with a buffered cursor:

1) The Cursor.executemany() method fails with error "InternalError:
No result set available" as described in the report.

2) Cursor.close() after Connector.cmd_query("SELECT 1") call
operation fails with Error "InternalError: Unread result found"
as described in the report.

With this patch, no fix was produced for issue 2 since it is
no longer happening, at least it could not be reproduced following
the report instructions. Testing was carried out on Windows and
Linux with Python versions 3.10.x and 3.8.x, respectively.

In the other hand, with this patch a fix is delivered for issue 1.
The CMySQLCursor.executemany() method was updated by following a
workflow similar to that of MySQLCursor.executemany() used by the
pure Python version.

Change-Id: I403bb946401bca9606eb287f02b9ac1e863539d4