BUG#33987119: TEXT and with a _bin collation (e.g: utf8mb4_bin) are considered as bytes object
An unexpected behaviour happens when using a "_bin" suffixed
collation type for the pure Connector/Python implementation.
BLOB and BINARY fields are expected to be delivered as bytes
objects as this is the expected behaviour, however, TEXT
fields are being delivered the same way when using a "_bin"
suffixed collation type. This latter is not correct.

With this patch, the unexpected behaviour is corrected by
updating the MySQLConverter._blob_to_python() method.
The connector at decoding/delivery time submits TEXT
fields to this latter method to determine the final conversion
since there is no specific method to handle TEXT datatypes.
An extra test is added, it is used the field charset description
to cast the incoming field value to the right type.

Change-Id: I5861bf6aedea7d4930ff9affb998fcf24a0d9ba8