WL#15036: Support for type hints
This worklog is the first effort of more to come, to incorporate type
hints over the Connector/Python codebase. With this edition, type hints
are added to methods and class attributes of files included in the
lib/mysql folder.

This worklog delivers a good starting annotations baseline, and these
can be improved gradually. To reach even higher annotations support,
a code refactory should be made.

Finally, a pre-commit hook was set for Mypy which is a static type checker
used to assess annotations' correctness.

Change-Id: I722f2df2a686437ed081558500bdb800107cfac7