BUG#34499578: MySQLCursor.executemany() fails to correctly identify BULK data loading ops
The MySQLCursor.executemany() method fails to batch insert data since
the regular expression (RE) sentinel used does not detect batch cases
correctly resulting in using the one-on-one insert, hence an overall
performance issue.

With this patch the RE sentinel is fixed, hence restoring back the
batch insert mechanism. Also, this patch fixes BUG#34555367 and is
the product of a contribution from Alex Cazacu.

Thank you for the contribution.

Change-Id: I784ea8c44d5f1f6dcd2d6d9448e173393689242d