BUG#34467201: Add init_command connection option
With this patch, support for the "inti_command" option is added as
part of the connection settings. This new option allows the user to
specify a command to be executed right after the connection is
established, that's to say, as part of the connection initialization.

Additionally, adding this option inside the Django database
configuration when connection is established now works,
previously this was ignored.

Thanks Sander van de Graaf for the contribution.

Change-Id: I24069cb071cdf388ebb543500c22db6d2c141d40