BUG#34675508: Character set 'utf8' unsupported in python mysql connector when using MariaDB
Connection fails when connecting to a MariaDB server using the MySQL
Connector/Python driver for adapter versions 8.0.30 and 8.0.31
due to unsupported charset. This issue is invariant to the type
of connection, that's to say, it happens when using a
pooled connection or not.

The bug happens because the default MySQL character collection,
that is 8.0 specific, does not recognize the utf8 charset name.
With this patch, the bug is fixed by using the LTS (5.7) MySQL
character collection as default instead. The new default will
switch to the 8.0 specific character collection when the queried
server version is 8.0.X.

Change-Id: I5b38bf27887ed25ae1d2d18dcd22637db260ae47