BUG#34655520: Wrong cursor.statement values in the results of cursor.execute(..., multi=True)
Wrong MySQLCursor.statement values are returned in the results of
cursor.execute(query_text, multi=True) when the following
conditions stand:

a) query_text is a string containing two queries or more,
separated by semicolons (;). For example,
query_text = "SELECT 1; SELECT 2".

b) The second query (or third, fourth, ...) uses a literal
or an identifier containing an odd number of backticks (`),
an odd number of quotes (') or an odd number of double quotes (").
For example, query_text = 'SELECT 1; SELECT "`"'

The expected answer is that each cursor from the result returns
one of the queries embedded in query_text when the corresponding
statement property is consulted. However, when the above conditions
stand, the before mentioned property may return more than one query.

With this patch, the behaviour is corrected by updating the regex
used to split out the executed operation (query_text).

Change-Id: I894509a0545947559818e0dc3ad588841b88eae1