BUG#23339387: Add MySQLCursorPreparedDict option
This patch integrates a new MySQLCursorPreparedDict class option that
is a blend of features from MySQLCursorDict and MySQLCursorPrepared.

The MySQLCursorPreparedDict class is similar to MySQLCursorPrepared,
the difference is that the former returns a fetched row as a
dictionary where column names are used as keys while the latter
returns a row as a traditional record (tuple).

To call out for MySQLCursorPreparedDict you should set
flags `prepared` and `dictionary`.

Thanks, Luke Weber for the contribution.

Change-Id: Iee40b5241c235d7ab3998325be4cf7823dad480f