BUG#34727432: Fix Django datetime error when USE_TZ=True in settings
When using the MySQL Connector/Python Django library as the backend in
an application where time zone aware values are set to be used
(USE_TZ=True), the `python migrate.py makemigrations` command fails
with a ValueError exception because at some point Django assumes a
given datetime value is naive and makes it zone aware. However, if the
value is non-naive, an error is raised.

With this patch, the issue is fixed by guaranteeing the
DjangoMySQLConverter class returns naive datetime values.

Change-Id: I9a46331d66b1bc865c4113c2caa0ec4d424abc1a