WL#15435: Improve the logging system
Connector/Python logging system currently has a logger per module,
which is cumbersome for users if they want to disable or configure
logging handlers for Connector/Python.

This worklog adds two loggers to configure Connector/Python logging
system, one for the classic protocol (mysql.connector) and the other
for the XDevAPI (mysqlx) implementations.

In addition, a small re-factoring in the authentication plugins imports
has been made to keep consistency.

Change-Id: I64b934dbf3b427eabdacad26ad2a4a150714f7a8