TypeAlias is not supported by earlier python versions
TypeAlias can be accessed directly from the `typing` module for python
version 3.10 and later, however for earlier versions trying to import
TypeAlias will fail (ImportError). For versions 3.9 and earlier,
TypeAlias can be accessed via the `typing_extensions` module,
nevertheless, this module is not part of the built-in modules, it has
to be installed as a 3rd party package.

At this point, we are not considering adding `typing_extensions`
as part of the module dependency list since it is not needed to run
the code, it is just needed by the type hint annotations.

As a consequence, a robust importing mechanism is implemented for
TypeAlias. First, we check if TypeAlias can be accessed from `typing`,
if not then we try to access it from `typing_extensions`,
if such a module is not installed we fallback to a workaround where
TypeAlias is set to be an alias of `typing.Any`.

The fallback option guarantees code execution goes on no matter if
TypeAlias can be imported or not. The downside, in case TypeAlias is
set to be `Any`, is that some variable types will not be inferred by
mypy since they will be treated as type `Any`.

Change-Id: I3d57172867c64eb08e00391d723a81e102d01489