BUG#27359063: Support for dictionary, named_tuple, and raw to prepared statements cursor
This bug report is an enhancement request regarding the available
prepared cursor flavors.

With this patch, two more options are added to the set of available
cursor types:
- `prepared + named_tuple`
- `prepared + raw`

The option `prepared + dictionary` was not added given it was
implemented while fixing BUG#23339387.

The three cursor types mentioned are available for both implementations;
pure python and cext.

Change-Id: Idd6276c1ac021ba759be6d5c5e93f18f1f41ed05