WL#15528: Add docker build/test support for existing server
This worklog introduces a Docker workflow for testing. The core
purpose of this new feature is to make unit testing more accessible,
faster, and more consistent by creating a base docker image where
the Connector/Python tests can be executed.

This workflow lightens the developers' burden of installing and
configuring dependencies to build the environment needed to
run the test suite, making contributions to the project
easier.

There is a script that builds and runs a Docker container which
then executes the test suite with the C extension enabled for both
packages `connector.mysql` and `mysqlx`. This means no external
dependency, apart from a running MySQL server, is needed from the
user's standpoint.

Change-Id: I41d53e553feeeba6911429b302bf65ef496d53eb