BUG#21476351: Too small chunks when sending huge parameters with COM_STMT_SEND_LONG_DATA
The bug could not be reproduced entirely since the pure python
implementation performed similarly to the C extension. However,
we took into consideration the idea of changing the number of
bytes to read (chunk size) when working with file-like objects
(e.g., io.BytesIO).

The chunk size is increased from 8 KB to 128 KB obtaining a
slight increase in performance, especially for bigger inputs.

Change-Id: Ieb2f5a36105b722d16c21c8eaebe501dd787e921