BUG#35015758: COM_QUIT should not be called in the connection phase
Connector/Python will call CMD_QUIT, close the socket connection,
and raise the right exception if something goes wrong in the
connection phase. However, by calling CMD_QUIT in this phase,
the server will interpret it as a "Bad Message" as the connection
never made it to the command phase, hence CMD_QUIT is an invalid move.

With this patch, CMD_QUIT is not called if something goes wrong in
the connection phase.

Change-Id: I845b21ed5f14911e40bdd1a106291acbafc619e4