WL#15591: Improve the network module
This work-log introduces an improvement to the network module
in the following regard:

1. Design
2. Implementation

Now the MySQLSocket interface isn't responsible for processing
`send` and `receive` events, this task is delegated to a
lower-level object named `Network Broker`. The latter change
allows the new design to be easier to comprehend and maintain.

The `Network Broker` is a component specifically designed to manage
communication events. Such a component is decoupled from the MySQLSocket
component, making it easy to customize and maintain as well.

This improvement also fixes BUG#29115406. The `recv` method
when working in compressed mode was incomplete since it did not handle
the case in which the last packet after unzipping a compressed payload
is incomplete.

Change-Id: Icb666a304082e32b083d1f322f5cfe1fc2c4cf4d