BUG#34984850: Fix binary conversion with NO_BACKSLASH_ESCAPES mode
The Connector/Python C extension converter truncates bytes containing
the \x00 byte when using NO_BACKSLASH_ESCAPES mode in MySQL.
The PyBytes_AsString() Python function interprets the \x00 as the end
of the string, dropping the rest of the binary data.

This patch fixes this issue by using the PyBytes_Concat() function to
quote the resulting bytes.

Change-Id: I3c17b9e6b52df10c2bdf654d085a08e171f87c65