WL#15629: Add OpenTelemetry tracing
With this worklog, OTel (OpenTelemetry) tracing support is added.
OTel already provides an `opentelemetry.MySQLInstrumentor`
supporting mysql-connector-python which is compatible with this
new feature, however, a new dedicated MySQL instrumentor is also
provided as part of this new feature.

This new instrumentor, which can be found in
`mysql.connector.opentelemetry`, exposes the same API and has
usage similar to `opentelemetry.MySQLInstrumentor`, nevertheless,
the trace morphology is different.

Additionally, a new connection property named
`otel_context_propagation` is added, and it can be used to control
the trace context propagation behavior. A value of `True` is used
as the default meaning that the trace context in progress (if any)
will be propagated to the server. If the user DOES NOT want the
propagation to happen no matter what, `otel_context_propagation`
should be set to `False`.

Change-Id: I0203924344ec6ad555fcf02531dd65b59c5a5193