BUG#35425076: Fix deallocating None error
The patch for BUG#35233031, which removes the default authentication
plugin mysql_native_password as the default replaces the variable type
in the C extension to Py_None without incrementing the reference count,
leading to a deallocation error when calling the connection close
multiple times.

This patch fixes the issue by setting the default value of the
authentication plugin to PyUnicode_FromString("") instead of Py_None.

Change-Id: I29314517fc55051d293f939c4e65ae58678e725a