BUG#29115406: CONTRIBUTION - FIX RECV COMPRESS BUG
When switching to compressed mode while using the pure python
implementation, there is a bug where a compressed packet may
contain an incomplete MySQL packet at last.

The bug report does not say much about the issue and does
not even provide a test case that proves it, however,
we managed to find the bug the contributor means while
developing the WL#15591 (improve the network module) where
the low-level network communication code was refactored.

This patch updates the test case previously implemented
in WL#15591 to check the issue so that the test payload
sizes do not surpass the default configuration of
`max_allowed_packet` for a common MySQL server setup.

Thanks, Allen Long for the contribution.

Change-Id: I99a90538553b49aa7c0737c7999cbe5b281d5fd3