BUG#35503377: First connected to server v8, then any v5 connections fail with utf8mb4 charset
If multiple connections are simultaneously open, they share
the same character set info [1], this works when using
parallelly a set of v8 [2] connections or a group of v5 [3]
connections, however, mixing them up might lead to an
error - for instance, the bug ticket reports a case when
two connections v8 and v5 are opened (in the mentioned order)
and their lifespans overlap.

[1] The class `CharacterSet` is defined as a static
entity - from properties down to methods.
[2]: Connections made to a MySQL server whose version >= 8.x.y
[3]: Connections made to a MySQL server whose version == 5.x.y

Change-Id: I11b584a2bd6c3a2d4b1c747ea9a6a7352baa4bcf