BUG#35755852: Django config raise_on_warnings is ignored without isolation_level
When the Django config option `isolation_level` is not provided, the
value of the option `raise_on_warnings` is ignored.

This issue was fixed by adding the missing default value in the pop()
function used for the `isolation_level`, which raises a `KeyError` and
breaks the assignment of the options.

Change-Id: I54a676b21586b94015618d16c223a60b5871221c