WL#15950: Support query parameters for prepared statements
With this WL, support for query attributes for prepared
statements is added, in particular for the C-EXT since
pure Python already supports it.

Additionally, we take this implementation as an excuse
to enable OpenTelemetry trace context propagation when
working with prepared statements in the C-EXT. Previously
it wasn't possible for the C-EXT since support for query
attributes for prepared statements was missing, but now
context propagation for this kind of statements is supported.

Change-Id: Ic119b05765788d502f26eb2feaec492fa788edfd