BUG#35710145: Bad MySQLCursor.statement and result when query text contains code comments
Wrong query results provided by the connector, when
the query contains code comments while using
`cursor.execute(query, multi=True)`. The regex expression
that splits a multi-statement query produces as many
statements as semicolon-separated operations are found in the query.

End-of-line (EOL) comments are treated as if they were
normal statements by the connector but that leads to
inconsistencies when reading results from the server
and trying to map each result to a statement since it's
assumed every statement will produce a result, but isn't
always the case - e.g. for comments.

With this patch, comments are simply ignored to guarantee
results and executed statements are matched correctly.

Change-Id: I728222d626bb562d23066e38924e8007faff5bc1