BUG#21390859: STATEMENTS GET OUT OF SYNCH WITH RESULT SETS
When including callproc (calling stored procedures)
statements as part of a multi-query string while
executing `cursor.execute(multi_stmt_query, multi=True)`
the statements get out of sync with the result, that's to say,
the statement-result mapping isn't correct, especially
when a procedure produces multiple results.

This issue is fixed as part of BUG#35710145, however,
no relevant tests for this use case were added during
the fix of BUG#35710145.

With this patch, we extend the unit tests to cover the
scenario where multi-result `callproc` statements are
included in a multi-statement query.

Change-Id: I30917e9ced90e5d0978929a714b295115c7083ab